fix: change path to moved interface
